requirements.txt
opencv-python
numpy
ultralytics
pillow
filterpy
scikit-image
lap
